feat: Fix URL frontmatter for indexes, add beta doctype cascade
This documentation commit makes two specific changes:

1. It adds the URL directive for folder indexes
2. It adds the beta doctype to all codes using cascade in the root index

Together, this PR ensures the correct, full URL for the product name
(NGINX Management Suite) is part of the path for ADM documentation, and
that each individual page in the documentation set has a warning that
the information included is in a beta state.

There are also some minor changes to section and page headings included,
and re-ordering of frontmatter parameters for consistency.