fix: Make core dump requirements explicit, fix shell command formatting
The previous version of this page suggested that you could enable a core
dump one of two ways: both methods are actually required to in order to
get core dumps to work.

Although outside of scope, I also fixed the formatting of a number of
codeblocks - explicitly indicating they were shell commands, and
removing the prepended prompt, which is not standardised across
environments.