feat: Add section & redirects for NGINX Ingress Controller
As part of the NGINX Ingress Controller 3.2 release, I have been working
on improving aspects of the troubleshooting documentation. Part of these
improvements includes renaming files in an effort to make them
consistent with their URL slugs and the intended content of the page.

This commit amends these changes to the existing redirects files, and
gives NGINX Ingress Controller its own distinct section, which I have
moved the single existing entry for the product into.