Add closing tag for bootstrap-table (#364)
### Proposed changes

The closing tag for the bootstrap-table shortcode was missing. I added
the closing tag to resolve the build failure
[here](https://app.netlify.com/sites/docs-nginx-com/deploys/64cc37aeda73a52ca46af2f9#L68),
which is related to the Hugo and theme upgrade.


### Checklist

Before creating a PR, run through this checklist and mark each step as
complete.

- [x] I have read the
[`CONTRIBUTING`](https://github.com/nginxinc/docs/blob/main/CONTRIBUTING.md)
document
- [x] I have viewed my changes in the deploy-preview
- [ ] I have updated any relevant supporting documentation
([`README.md`](https://github.com/nginxinc/docs/blob/main/README.md) and
the
[`CHANGELOG.md`](https://github.com/nginxinc/docs/blob/main/CHANGELOG.md))

When you're ready to merge a PR, run through this checklist and mark
each step as complete.

- [ ] If you moved or renamed a publicly-accessible doc file, add the
corresponding redirect to the
[_redirects](https://github.com/nginxinc/docs/blob/main/_redirects),
[_redirect_dev](https://github.com/nginxinc/docs/blob/main/_redirects_dev),
and
[_redirects_staging](https://github.com/nginxinc/docs/blob/main/_redirects_staging)
files.
- [ ] Review the doc for spelling errors.
- [ ] Verify that all links in the doc work.
- [ ] Verify that the doc follows the appropriate content template.
- [x] Add `@nginxinc/nginx-docs` as doc reviewers.
- [x] Share the PR for review in the `#docops-mr-reviews` channel in
Slack.