feat: Rewrite sections, adjust information architecture and rename file
This commit adds a large number of changes, starting first by renaming
the actual file (which affects the default URL slug in production). From
the top down I rewrote many sentences and paragraphs in an attempt to
make them more concise.

The information architecture was updated to add clear delimiters for
each section alongside rephrasing many of the headings to better fit
title case and overall flow for the page.

Finally, there are changes made to the final sections of the page to
conform to the API/UI tabbed instructions pattern commonly seen
elsewhere in docs (Such as in ACM policies), but this is a placeholder
as my lack of familiarity has lead me to defer to the original author of
the document to fill in these sections.