feat: Change subheading formatting, re-add custom styles tag
This commit changes the subheading formatting for checking the instance
states to be closer to the normal convention for NGINX documentation, by
removing the enumeration, punctuation and shortening the length.

It also re-adds the custom styles tag which is necessary for the JSON
request example in the penultimate section, which has also had a slight
rephrasing and reformatting in its preceding paragraph.