fix: Remove .idea folder, update gitignore for wildcard folder matching
The .idea file which contains workspace information for Jetbrains'
IntelliJ IDEA IDE was erroneously checked into the repo. This commit
removes the folder, and also updates the base .gitignore file.

The change to the .gitignore file swaps some of the entries to wildcard
matches, so they will be ignored no matter what level of nesting they
might appear on the filesystem.