fix: Update wiki link to instead point to community awesome-nginx repo
The wiki has been archived for a long time. As part of ongoing efforts
to formally remove it, we are removing references to it when they are
found. This PR changes the link from NGINX Plus's dynamic modules page
to instead point to an active awesome-nginx repository managed by a
community member, which has accepted PRs from NGINX employees.