feat: Fix broken images in NGINX Controller documentation
As reported in issue #366, the "How to Use the Data Explorer" document
had broken images. By generating the public Hugo files I determined the
given filepath for the images was incorrect.

This commit fixes it by following the filepath pattern set using the img
shortcode elsewhere in the controller docset, though I believe the
pattern for this asset generation to be a bug.

Since it's from a deprecated product, it could just be an artefact that
was addressed with better design for other products.