Adds Documentation For NMS NAP Compiler 4.7 (#577)
### Proposed changes

This PR is responsible for updating the example NMS NAP Compiler
commands to use the newest NAP version, which at the time of this PR it
is 4.7.0. Additionally, the NIM NAP compatibility matrix is updated as
well.

**JIRA ID**: 
- N/A

**GitHub Issue**:
- N/A

### Checklist

Before creating a PR, run through this checklist and mark each step as
complete.

- [x] I have read the
[`CONTRIBUTING`](https://github.com/nginxinc/docs/blob/main/CONTRIBUTING.md)
document
- [ ] I have viewed my changes in the Netlify deploy-preview
- [x] My PR is targeting the correct branch:
  
  - main: content that can be released immediately
- product release branch: content that should be held for a future
release
    
- [x] I have updated any relevant supporting documentation
([`README.md`](https://github.com/nginxinc/docs/blob/main/README.md) and
the
[`CHANGELOG.md`](https://github.com/nginxinc/docs/blob/main/CHANGELOG.md))
- [x] I have followed the [conventional commits
guidelines](https://www.conventionalcommits.org/en/v1.0.0/#summary) for
all commits on my branch
- [x] I have updated the
[_redirects](https://github.com/nginxinc/docs/blob/main/_redirects),
[_redirect_dev](https://github.com/nginxinc/docs/blob/main/_redirects_dev),
and
[_redirects_staging](https://github.com/nginxinc/docs/blob/main/_redirects_staging)
files to reflect changes to any public-facing doc files that I moved or
renamed.

When you're ready to merge a PR, run through this checklist and mark
each step as complete.

- [ ] Review the doc for spelling errors.
- [ ] Verify that all links in the doc work.
- [ ] Verify that the doc follows the appropriate content template.
- [ ] Add technical and docs reviewers.
  - @nginxinc/docs: any PR (if you're not sure, choose this team)
  - @nginxinc/nap-docs-approvers: add for NGINX App Protect docs.
- @nginxinc/plus-docs-approvers: add for NGINX OSS and NGINX Plus docs.

- [ ] Share the PR for review in the `#docops-mr-reviews` channel in
Slack.