feat: Remove module references from nms.conf documentation page
This commit removes the references to ACM, ADM & NIM from the NGINX
Management Suite documentation for configuration using a file. Initially
I started rewriting the opening paragraph to reference only NIM
singularly instead of modules in plurality.

When I read the page and example YAML I realised there was no reference
to any of the modules beyond their mention at the top of the page, and
made the decision to remove them altogether as they seem to have been
dropped into the top of the page purely for visibility with no relevance
to the actual intent of the documentation.