feat: Add OpenShift warning to NIM K8s deploy instructions
This commit adds a warning call-out that OpenShift is not a supported
platform for NIM or (And by proxy, NMS as a whole). It also fixes the
page weights of the subsection to be more consistent.

There is more work to be done on this page, which will include moving
this warning somewhere better - I will have a subsequent PR related to
another ticket.

This closes #908.