feat: Update pull request template
We no longer use Netlify, and NGINX has migrated away from Netlify
to an internal solution for documentation previews.

The technical and docs reviewers guidance seems like a workaround for
what could be a robustly organised CODEOWNERS file, so I may have a
follow up PR dependent on the results of a brief investigation.