fix primaryIdlFileName in grpc-content-profiles.md for NAP (#1260)
Fix the typo, and make sure it is inline with
https://github.com/nginxinc/docs/blob/65903db65c1a5b9c7c503ba911ee2a45a45510e7/data/nap-waf/schema/policy.json#L1948

### Proposed changes

Describe the use case and detail of the change. 

If this PR addresses an [issue](https://github.com/nginxinc/docs/issues)
on GitHub, ensure that you link to it below.

**GitHub Issue**:
Closes #<Issue>
https://jira.f5net.com/browse/WAFMC-11853 

### Checklist

Before creating a PR, run through this checklist and mark each step as
complete.

- [x] I have read the
[`CONTRIBUTING`](https://github.com/nginxinc/docs/blob/main/CONTRIBUTING.md)
document
- [x] I have viewed my changes in the documentation Deploy Preview
- [x] My PR is targeting the correct branch:
  - main: content that can be released immediately
- product release branch: content that should be held for a future
release
- [x] I have updated any relevant supporting documentation
([`README.md`](https://github.com/nginxinc/docs/blob/main/README.md) and
the
[`CHANGELOG.md`](https://github.com/nginxinc/docs/blob/main/CHANGELOG.md))
- [ ] I have followed the [conventional commits
guidelines](https://www.conventionalcommits.org/en/v1.0.0/#summary) for
all commits on my branch

When you're ready to merge a PR, run through this checklist and mark
each step as complete.

- [x] Review the doc for spelling errors.
- [x] Verify that all links in the doc work.
- [x] Verify that the doc follows the appropriate content template.
- [ ] Add technical and docs reviewers.
- [ ] Share the PR for review in the
[`#nginx-doc-reviews`](https://f5.enterprise.slack.com/archives/C04PYFULN91)
channel in Slack.