feat: Update NGF Logo in website layout (#1479)
NGINX Gateway Fabric is now using the old NGINX Service Mesh logo: this
PR makes the change on the documentation website side. A similar PR will
be created within the Gateway Fabric repo to update their configuration.

The logo file is being sourced from the theme, which was updated with
the images previously.