feat: Update Security Monitoring section, fix NAP link
This commit updates the Security Monitoring section to fit contemporary
style standards: sentence case for titles, no gerunds, horizontal line
breaks at the end of sections.

It also fixes the frontmatter for the section's pages to be uniform in
order and parameters with newer pages, and adds explicit weights for
page numbering.

The final change is to add a reference link to the right NAP V4 page:
prior to the release of NAP V5 the link was valid, but was not
automatically detected due to being a production URL.