feat: Remove duplicate includes folder, update frontmatter
This commit removes an accidental duplicate of the includes folder, and
updates the frontmatter of every single page in the documentation set.
They are now all consistent in their fields and formatting, and a
handful of IA or heading changes have been made as low-hanging fruit to
bring the documentation set closer to full consistency.