feat: Add explicit URL parameter to Gateway Fabric index pages
This commit adds an explicit URL to the subfolders of the NGINX Gateway
Fabric set, which is inherited by the pages within the folders. It is
necessary for URLs to resolve sensibly, as otherwise they default to the
folder path, /ngf/, which we do not map as part of the deployment.