feat: Update PR links to GitHub URLs, fix inline Tutorial guidance
This commit updates the links from the PR templates to use the actual
GitHub links: they are currently relative links that do not parse as
intended, instead linking to parts of the GitHub website.

It also fixes an issue in the Tutorial archetype's inline Markdown
guidance caused by an unintentionally escaped quotation mark.