Use policy json instead of tgz
Using polciy JSON everywhere as our code will translate
it to tgz post complilation anyway. The service
accepts JSON from the customer as it is easier to
read and updating our docs to just use json in the
configuration just streamlines the experience. If
the user finds out that they can use tgz as well,
:shrug: