feat: Update Contributing guidelines for writers
This commit updates the CONTRIBUTING_DOCS.md file to our contemporary
standards for documentation formatting, adding mention of Hugo
archetypes and our ghcode shortcode, while also removing some slightly
redundant information.

Certain sentences have been removed or simplified to reduce the amount
of contextual reading required: in certain cases, the reader is guided
to refer to the style guide for more precise information, such as in the
case of using images.

There will likely be follow-up work in the future to more cleanly
delineate guidance for using Hugo and guidance around writing content in
a general sense, to tightly scope the purpose of this document in
relation to the style guide.