feat: Add prefix to custom frontmatter in Hugo archetypes
To reduce name collisions and make it explicitly clear which frontmatter
fields are used by DocOps (As opposed to native Hugo fields which
interact with specific mechanisms), this PR adds an `nd` prefix.

This ensures that documents generated using `hugo new` will have the
correct fields in the future. Follow up work to update existing fields
and documents will be handled in other pull requests.