feat: Update NGINX Gateway Fabric information architecture (#600)
This commit partially restructures the NGINX Gateway Fabric documentation
to flatten the folder structure, reducing the amount of steps a reader must 
navigate to get to useful instructions.

The majority of the effort is concentrated on the Install section, simplifying 
the URL structure and creating logical links in the sequence of pages presented 
and the linear connection of pre-requisites and next steps when a user is onboarding a cluster.

Additional work will follow this effort, once identifying more common user 
journeys and how individual documents can be grouped together logically.

---------

Co-authored-by: Saylor Berman <s.berman@f5.com>