feat: Clarify 3.x to 4.x upgrade process for NIC
This commit renames the Upgrade to 3.x document for NGINX Ingress
Controller to clarify it is for 3.x to 4.00 only. It adds an explicit
warning call-out at the top of the page to reinforce this, and the Helm
installation document has also been updated with more generic phrasing.

At a certain point of maturity the note on the Helm instructions page
could likely be removed, unless we anticipate someone will be still be
running a 3.x instance years after the versions have moved on.