feat: Clarify NIC install & upgrade steps
This commit makes a number of changes to clarify what the precise steps
are to install, upgrade and uninstall NGINX Ingress Controller

- The Install document clarifies necessary steps
- Uninstall steps have been moved to the bottom of the document
- Upgrade instructions have been moved into their own document

The changes made focused on clarifying the critical path for one single
user story, as lots of unnecessary contextual information was being
provided, which was causing confusion.

With this commit, the general reading and execution flow for these
deployment steps become more coherent. Similar changes may be made for
Manifests documentation, if common steps are identified.