feat: Update old call-outs to use new style
This commit replaces all instances of the old call-out style with the
new one. The new style is more generic, with contributors using a
parameter to denote the type of call-out.

Instances of the new call-out style are also more consistent to our
general Hugo formatting style: previously, empty space usage was highly
inconsistent.

These changes should address the terminal warnings, and are more
futureproof for updates to the theme in the future.