feat: Add NGINX Ingress Controller landing page
This commit adds a custom landing page for NGINX Ingress Controller
using a relatively new landing page archetype. The landing page is
intended to give an impression of the product and provide links to some
of the more commonly visited pages of both new and returning readers.

The addition of these landing pages is a requirement for the new
documentation theme's release, which receives ongoing updates.