Adds a callout to NIC installation with helm
We ran into an issue where helm version 3.18.5 is broken because NIC
charts reference definitions using an https URL. Helm rejects any
reference that's not a file:// scheme, so linting fails.

The helm team is aware of the bug and a fix is expected in the next
week. Next release should work as previously.