Updates nginxaas-google overview and architecture
Updated the page using nginxaas-azure as a template. The text that
existed was describing a diagram that is not appropriate to display to
customers. This has been replaced with bullet points I think will match
the eventual marketing diagram. The bullet points were sanitized of
references or links to Azure and include a little detail about
networking via Private Service Connect (PSC).

Under **Redundancy** I replaced the name of the plan with "Enterprise"
as this is what we've called the single plan that we currently have. We
do not expect there to be another plan for other "service levels" (e.g.,
a basic or development plan). These service levels do not yet exist in
NGINXaaS for Google Cloud.

There was a note about redundancy features which I replaced with a
mention of utilizing "multiple zones within each region". I'm not sure
anything else needs to be said about that here.