Change timeline for NGINX updates on stable channel
We want to be able to deliver the latest NGINX release sooner. Currently we state the minimum delay between an. NGINX release and when NGINXaaS will deploy it to the stable channel is 45 days.  We want to shorten it to 24 days (this is a minimum and doesn't require that we release in 24 days - we could take longer depending on the significance of the changes).