docs: fix n4g filesystem restrictions table
The nginx master process, like the worker process, can write to /spool/nginx and /var/spool/nginx.
This also removes the /var/log/nginx row which we don't want to advertise as scratch space for
the user because the contents of that directory can get exported to google cloud logging.