feat: Replace details HTML with Hugo shortcode (#1321)
This commit replaces the majority of instances of raw details HTML 
in the documentation repo with the Hugo shortcode instead. 

The remaining instances are exceptions to our rule of avoiding 
raw HTML in documentation, as they are bespoke interactive elements.

The affected files have also had other formatting styles brought up 
to contemporary standards, such as metadata, the deprecation of 
the bootstrap shortcode, and the removal of horizontal lines.