feat: Convert important psuedo-callouts
This commit converts the majority of important psuedo-callouts in
documentation to proper Hugo call-out shortcodes. The instances that
have been left are parts of files that will be deprecated in the future.

As part of this commit, metadata and Markdown linting issues were also
fixed, and some peripheral process documentation was updated for
clarity.