feat: Add deprecated content process documentation
This commit adds a process document for reviewing old, deprecated
content. This was always possible thanks to git revision history, but
there was no single reference point available for when deprecated
content was removed.

This new process document is a quick reference point, allowing someone
to check out the relevant state of documentation for a given product
without needing to undergo any git archaeology.