feat: Add pseudo-variables to F5 WAF for NGINX
This commit uses the shortcode psuedo-variable pattern to make multiple
component versions for F5 WAF for NGINX into strings. This allows for
content in multiple locations to be changed at once by updating the file
relating to the version.

Many of the versions are currently identical: if the individual
components all use the same version, then the amount of shortcode files
can be reduced accordingly.

- Closes #1318 by replacing the string with a shortcode
- Closes #1364 by creating shortcode psuedovariables for each item