feat: Consolidate Glossary pages (#1374)
This commit consolidates the various Glossary pages across the 
NGINX product documentation sets into a single page with its own
card on the documentation landing page. This was largely accomplished
by making the contents of each page into an include file, so the
definitions can be managed in a single place while being rendered in
the original product sets and the unified Glossary page.

---------

Co-authored-by: Mike Jang <3287976+mjang@users.noreply.github.com>
Co-authored-by: Alan Dooley <github@adubhlaoi.ch>