Update naas docs to call out expectations on network attachments (#1429)
We have cases where users will have network
attachments with a list of projects allowed on it.
We are making changes to the service to handle
this better and float the error to the user.

This change calls out what the user sees and
should do in a situation where  flow they have
project allowlists on the network attachment.