Update NGF service docs again
Problem: The Service docs were determined to be in the wrong place based on some changes in the docs structure with NGF v2.

Solution: Move the Service docs to a more relevant section where the NGINX Service is actually created. Also removed some extraneous info that I didn't think was necessary.