feat: Add user browsers and 2023 changelog content for WAF
This commit adds the "User-defined browser control" feature page to the
policy section of F5 WAF for NGINX documentation, and changelog entries
for 2024 and 2023 before the v4/v5 split.

It also adds a notice to the top of the deny/allow IP feature page and
supported feature table directing users to use the IP addresses feature
instead, which is newer and deprecates the latter.