feat: Rewrite Microsoft Azure documentation for NGINX (#1507)
This commit builds atop a community PR that has been inactive for
months, rewriting the documentation for creating a Microsoft Azure
virtual machine that uses NGINX. The changes include:

- Removing redundant text and indentation
- Delineating NGINX Plus information with call-outs
- Replacing inline HTML with Hugo shortcodes

There are additional improvements that could be made, such as avoiding
re-documenting a third party SaaS altogether. The scope of such a change
is outside the remit of this PR, as it would be an overall content
design change for how NGINX documentation is structured and managed.

---------

Co-authored-by: jawfu-M <Jawfumaryam@gmail.com>
Co-authored-by: Mike Jang <3287976+mjang@users.noreply.github.com>
Co-authored-by: yar <y82@users.noreply.github.com>
Co-authored-by: Jon Torre <78599298+JTorreG@users.noreply.github.com>