feat: Add version instructions for WAF virtual environments
This commit adds instructions to each platform in the F5 WAF for NGINX
virtual environment installation document to explain how to find and
install specific package versions.

As the information is optional, it is hidden by default using a details
element, and introduced at the point where the user would install a
package, as all other steps in the process are identical.