NGINXaaS: Address missing info and points of user friction (#1588)
* naas: Add format of network attachment

This makes it easier for the user to know what it
looks like when we are asking them to save it for
use later on in the deployment.

* Call out where a user can find their deployment service account ID

This is important as we do not have references to
it anywhere and is information needed by the user
to set up deployment observability.

* Link users to monitoring docs from IAM prereqs

After I am done setting up my WIF provider, I
should be directed to enable monitoring and
logging for my deployment.

* Simplify intro to WIF a bit

* Improve warning around setting up network attachments

We want to strongly recommend using a connection
preference that is secure but keep the
instructions open for a user that may not be as
concerned with security or just wants to
prioritize deploying quickly just to try out the
service.