feat: Improve F5 WAF multi-container Docker deployment (#1594)
This commit updates the instructions for an F5 WAF for NGINX
multi-container Docker improvement to ensure that the mounts are
consistent across containers and that an address is given a formal
placeholder, indicating it must be replaced.

This closes #1488, which also had all of the detail necessary for the
fix thanks to @mrajagopal's level of detail in the issue. There is extra
detail in the issue that could be used for further improvement of this
use case in the future.